+++ backend/canisters/notifications_index/impl/src/updates/remove_subscriptions.rs
@@ -13,6 +13,9 @@ fn remove_subscriptions(args: Args) -> Response {
 fn remove_subscriptions_impl(args: Args, state: &mut RuntimeState) -> Response {
     let now = state.env.now();
     for user in args.subscriptions_by_user {
+        for endpoint in user.endpoints {
+            state.remove_subscription(user.user_id, endpoint, now);
+        }
         for key in user.p256dh_keys {
             state.remove_subscription(user.user_id, key, now);
         }