+++ frontend/app/src/components/App.svelte
@@ -19,7 +19,6 @@
     } from "@utils/native/notification_channels";
     import "@utils/scream";
     import {
-        isCanisterUrl,
         isLandingPageRoute,
         isScrollingRoute,
         redirectLandingPageLinksIfNecessary,
@@ -53,7 +52,6 @@
     import Head from "./Head.svelte";
     import Profiler from "./Profiler.svelte";
     import Snow from "./Snow.svelte";
-    import SwitchDomain from "./SwitchDomain.svelte";
     import UpgradeBanner from "./UpgradeBanner.svelte";
     import Witch from "./Witch.svelte";
     import InstallPrompt from "./home/InstallPrompt.svelte";
@@ -61,7 +59,6 @@
     import ActiveCall from "./home/video/ActiveCall.svelte";
     import IncomingCall from "./home/video/IncomingCall.svelte";
     import VideoCallAccessRequests from "./home/video/VideoCallAccessRequests.svelte";
-    import Upgrading from "./upgrading/Upgrading.svelte";
 
     overrideItemIdKeyNameBeforeInitialisingDndZones("_id");
 
@@ -132,10 +129,7 @@
     let burstUrl = $derived(isFirefox ? `${burstPath}.png` : `${burstPath}.svg`);
     let burstFixed = $derived(isScrollingRoute($routeStore));
 
-    let upgrading = $derived(
-        $identityStateStore.kind === "upgrading_user" ||
-            $identityStateStore.kind === "upgrade_user",
-    );
+    let lastScrollY = $state(window.scrollY);
 
     trackedEffect("rtl", () => {
         // subscribe to the rtl store so that we can set the overall page direction at the right time
@@ -638,11 +632,7 @@
 
 <NotificationsBar />
 
-{#if isCanisterUrl}
-    <SwitchDomain />
-{:else if upgrading}
-    <Upgrading />
-{:else if $identityStateStore.kind === "anon" || $identityStateStore.kind === "logging_in" || $identityStateStore.kind === "registering" || $identityStateStore.kind === "logged_in" || $identityStateStore.kind === "loading_user" || $identityStateStore.kind === "challenging"}
+{#if $identityStateStore.kind === "anon" || $identityStateStore.kind === "logging_in" || $identityStateStore.kind === "registering" || $identityStateStore.kind === "logged_in" || $identityStateStore.kind === "loading_user" || $identityStateStore.kind === "challenging"}
     {#if !$isLoading || $reviewingTranslations}
         <Router {showLandingPage} />
     {/if}