+++ frontend/openchat-client/src/openchat.ts
@@ -5214,8 +5214,8 @@ export class OpenChat {
         return this.config.i18nFormatter("unknownUser");
     }
 
-    #subscriptionExists(p256dh_key: string): Promise<boolean> {
-        return this.#sendRequest({ kind: "subscriptionExists", p256dh_key }).catch(() => false);
+    #subscriptionExists(endpoint: string, p256dh_key: string): Promise<boolean> {
+        return this.#sendRequest({ kind: "subscriptionExists", endpoint, p256dh_key }).catch(() => false);
     }
 
     #pushSubscription(subscription: PushSubscriptionJSON): Promise<void> {
@@ -9921,7 +9921,7 @@ export class OpenChat {
         if (pushSubscription) {
             console.debug("PUSH: found existing push subscription");
             // Check if the subscription has already been pushed to the notifications canister
-            if (await this.#subscriptionExists(this.#extract_p256dh_key(pushSubscription))) {
+            if (await this.#subscriptionExists(pushSubscription.endpoint, this.#extract_p256dh_key(pushSubscription))) {
                 console.debug("PUSH: subscription exists in the backend");
                 return true;
             }
@@ -9989,7 +9989,7 @@ export class OpenChat {
         if (registration !== undefined) {
             const pushSubscription = await registration.pushManager.getSubscription();
             if (pushSubscription) {
-                if (await this.#subscriptionExists(this.#extract_p256dh_key(pushSubscription))) {
+                if (await this.#subscriptionExists(pushSubscription.endpoint, this.#extract_p256dh_key(pushSubscription))) {
                     console.debug("PUSH: removing push subscription");
                     await this.#removeSubscription(pushSubscription.toJSON());
                 }